prep spec_helper.rb for PDK upgrade
