+++ .sync.yml
@@ -32,3 +32,14 @@ spec/default_facts.yml:
       - 2.5.0
     vm_guests:
       - invalid_existing_guest
+spec/spec_helper.rb:
+  spec_overrides:
+    - "module DefaultHieraConfig"
+    - "  extend RSpec::SharedContext"
+    - "  let(:hiera_config) { 'spec/fixtures/hiera/default_config.yaml' }"
+    - "end"
+    - ""
+    - "RSpec.configure do |c|"
+    - "  c.default_facts = default_facts"
+    - "  c.include DefaultHieraConfig"
+    - "end " # trailing space prevents PDK from skipping this line
