+++ spec/spec_helper_local.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+require 'faker'
+require 'spec_helper/deep_merge'
+require 'spec_helper/stub_loader'
